Migrates test cases for ComputeCurrentTime
